move invoice table rows into a separate component
