+++ app/components/billing-invoice-list/template.emblem
@@ -1,0 +1,12 @@
+td #{item.id}
+td {{moment-format item.created "DD MMM YYYY"}}
+td {{format-currency item.currency item.amount}}
+td #{item.status}
+td class="has-text-centered"
+  button class="is-primary" click={(perform downloadInvoice item)} disabled=downloadInvoice.isRunning
+    if downloadInvoice.isRunning
+      i class="fa fa-spin fa-spinner"
+    else
+      i class="fa fa-download"
+    | &nbsp;&nbsp;
+    t "invoiceTable.actionButton.download"
