+++ tests/integration/components/billing-invoice-list/component-test.js
@@ -1,0 +1,24 @@
+// import { moduleForComponent, test } from 'ember-qunit';
+// import hbs from 'htmlbars-inline-precompile';
+
+// moduleForComponent('billing-invoice-list', 'Integration | Component | billing invoice list', {
+//   integration: true
+// });
+
+// test('it renders', function(assert) {
+//   // Set any properties with this.set('myProperty', 'value');
+//   // Handle any actions with this.on('myAction', function(val) { ... });
+
+//   this.render(hbs`{{billing-invoice-list}}`);
+
+//   assert.equal(this.$().text().trim(), '');
+
+//   // Template block usage:
+//   this.render(hbs`
+//     {{#billing-invoice-list}}
+//       template block text
+//     {{/billing-invoice-list}}
+//   `);
+
+//   assert.equal(this.$().text().trim(), 'template block text');
+// });
